+++ txn_test.go
@@ -17,6 +17,9 @@
 package aerospike_test
 
 import (
+	"fmt"
+	"time"
+
 	as "github.com/aerospike/aerospike-client-go/v8"
 	"github.com/aerospike/aerospike-client-go/v8/types"
 
@@ -53,7 +56,20 @@ var _ = gg.Describe("Aerospike", func() {
 				function writeBin(r,name,value)
 					putBin(r,name,value)
 				end
-			`
+
+				function get_gen(rec)
+					return record.gen(rec)
+				end
+				
+				function rec_read(rec)
+					local m = map()
+					names = record.bin_names(rec)
+					for i, bn in ipairs(names) do
+						m[bn] = rec[bn]
+					end
+					return m
+				end
+				`
 
 			regTask, err := client.RegisterUDF(nil, []byte(luaFunc), "record_example.lua", as.LUA)
 			gm.Expect(err).ToNot(gm.HaveOccurred())
@@ -632,6 +648,99 @@ var _ = gg.Describe("Aerospike", func() {
 				gm.Expect(records[i]).ToNot(gm.BeNil())
 				gm.Expect(records[i].Bins[binName]).To(gm.Equal(1))
 			}
+		}) // it
+
+		gg.It("UDF should not read expired record", func() {
+			k, _ := as.NewKey("test", "demo", 0)
+			client.PutBins(nil, k, as.NewBin("bin", 10))
+
+			txn := as.NewTxn()
+			wp := as.NewWritePolicy(0, 0)
+			wp.Txn = txn
+
+			client.PutBins(wp, k, as.NewBin("bin", 20))
+
+			time.Sleep(40 * time.Second)
+
+			p := as.NewPolicy()
+			p.Txn = txn
+
+			r, err := client.Get(p, k)
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(r.Bins["bin"]).To(gm.Equal(10))
+
+			r, err = client.Get(nil, k)
+			gm.Expect(r.Bins["bin"]).To(gm.Equal(10))
+
+			_, err = client.Commit(txn)
+			gm.Expect(err).To(gm.HaveOccurred())
+
+			p = as.NewPolicy()
+			p.Txn = txn
+			_, err = client.Get(p, k)
+			gm.Expect(err).To(gm.HaveOccurred())
+
+			wp = as.NewWritePolicy(0, 0)
+			wp.Txn = txn
+			_, err = client.Execute(wp, k, "mrt_ops", "rec_read")
+			gm.Expect(err).To(gm.HaveOccurred())
+		})
+
+		gg.It("must handle different key pointers", func() {
+
+			getGen := func(cli *as.Client, key *as.Key, txn *as.Txn) int {
+				var wp *as.WritePolicy
+				if txn != nil {
+					wp = as.NewWritePolicy(0, 0)
+					wp.Txn = txn
+				}
+				if val, err := cli.Execute(wp, key, "record_example", "get_gen"); err == nil {
+					return val.(int)
+				} else {
+					panic(err)
+				}
+			}
+
+			key, _ := as.NewKey(ns, set, []byte("0"))
+			err = client.PutBins(nil, key, as.NewBin("bin", 1))
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(getGen(client, key, nil)).To(gm.Equal(1))
+
+			txn := as.NewTxn()
+
+			wp1 := as.NewWritePolicy(0, 0)
+			wp1.Txn = txn
+			key, _ = as.NewKey(ns, set, []byte("0"))
+			err = client.PutBins(wp1, key, as.NewBin("bin", 2))
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(getGen(client, key, txn)).To(gm.Equal(2))
+
+			key, _ = as.NewKey(ns, set, []byte("0"))
+			err = client.PutBins(wp1, key, as.NewBin("bin", 2))
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(getGen(client, key, txn)).To(gm.Equal(3))
+
+			key, _ = as.NewKey(ns, set, []byte("0"))
+			err = client.PutBins(wp1, key, as.NewBin("bin", 2))
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(getGen(client, key, txn)).To(gm.Equal(4))
+
+			key, _ = as.NewKey(ns, set, []byte("0"))
+			err = client.PutBins(wp1, key, as.NewBin("bin", 2))
+			gm.Expect(err).ToNot(gm.HaveOccurred())
+			gm.Expect(getGen(client, key, txn)).To(gm.Equal(5))
+
+			key, _ = as.NewKey(ns, set, []byte("0"))
+			if err := client.PutBins(wp1, key, as.NewBin("bin", 2)); err != nil {
+				fmt.Println(err.Error())
+			}
+			gm.Expect(getGen(client, key, txn)).To(gm.Equal(6))
+
+			cs, err := client.Commit(txn)
+			gm.Expect(err).NotTo(gm.HaveOccurred())
+			gm.Expect(cs).To(gm.Equal(as.CommitStatusOK))
+
+			gm.Expect(getGen(client, key, nil)).To(gm.Equal(7))
 		}) // it
 	}) // describe
 })